Move esclient to gateway
